+++ src/dt.py
@@ -3,7 +3,7 @@ from transformers import (
     DecisionTransformerModel,
 )
 
-from _datasets import load_dataset
+from get_datasets import load_dataset
 
 class TrainableDT(DecisionTransformerModel):
     def __init__(self, config):
