+++ src/get_feedback.py
@@ -7,7 +7,7 @@ import numpy as np
 from numpy.random import default_rng
 
 from argparsing import get_feedback_args
-from _datasets import load_dataset, name_dataset
+from get_datasets import load_dataset, name_dataset
 
 OBJECT_TO_IDX = {
     "unseen": 0,
